Setup formatting for amount and createdAt

+++ src/components/ExpenseForm.js
@@ -60,7 +60,7 @@ class ExpenseForm extends React.Component {
             // console.log('Sumbitted')
             this.props.onSubmit({
                 description: this.state.description,
-                amount: parseInt(this.state.amount, 10) * 100,
+                amount: parseFloat(this.state.amount, 10) * 100,
                 createdAt: this.state.createdAt.valueOf(),
                 note: this.state.note
             })
